# api/openapi/payments-platform.v1.yaml
openapi: 3.1.0
...
info:
  title: Payments Platform API
  version: 1.0.0
  summary: Multi-tenant payment gateway onboarding and payment orchestration.
  description: |
    The public REST surface of the payments platform. Two products share this contract:

    * **Control plane** (`/v1/merchants`, `/v1/gateways`) — merchant registry, onboarding
      workflow, desired-state configuration and the gateway catalogue. Availability target
      99.9 %. Never on the payment hot path.
    * **Data plane** (`/v1/payments`, `/v1/webhooks`) — payment orchestration against
      third-party gateways. Availability target 99.99 %, p99 <= 250 ms excluding gateway time.

    ## Binding rules

    * **Money** is always an integer in the currency's minor units plus an ISO-4217 code:
      `{"amount": 1050, "currency": "USD"}`. Never a float, never a decimal string.
    * **Identifiers** are `<prefix>_<26-char Crockford Base32 ULID>` and are opaque to
      clients. Do not parse them.
    * **No cardholder data.** This API accepts gateway tokens and vault references only.
      Any request field containing a Luhn-valid 13-19 digit sequence is rejected with
      `400 SENSITIVE_DATA_IN_REQUEST` and the value is never logged. See
      `PaymentMethodReference`.
    * **Idempotency.** Every unsafe operation requires an `Idempotency-Key` header. Replays
      return the stored response with `Idempotent-Replay: true`. A key reused with a
...
      `If-Match`. A mismatch is `412`.
...
  contact:
...
  license:
...
servers:
...
security:
  - oauth2:
      - merchants:read
paths:
...
  /v1/payments:
    post:
      operationId: createPayment
      tags: [Payments]
      summary: Create a payment
      description: |
        Creates a payment and dispatches it to the highest-ranked eligible gateway.

        The request pipeline runs in this order and the order is load-bearing:
        authenticate, resolve tenant, authorize, rate limit, L1 schema validation including
        the PAN detector, idempotency claim, merchant context load, L5 payment validation,
        risk engine, routing, attempt persistence, gateway dispatch, L6 response validation,
        L7 state transition plus outbox write in one transaction, idempotency completion.

        **Timeout semantics.** If the gateway call times out or returns an ambiguous transport
        error, the attempt is recorded `TIMEOUT_UNKNOWN`, the payment **stays** `PROCESSING`
        and this endpoint returns `202` with `state: PROCESSING`. It does not return an error,
        because we do not know whether money moved. No timer may fail a payment. Resolution
        comes from a gateway webhook, a reconciler lookup keyed on our deterministic gateway
        idempotency key, or a settlement report. Poll `GET /v1/payments/{paymentId}`.

        **Never** re-issue a timed-out payment with a new idempotency key. Retry with the same
        key, or poll.
      security:
        - oauth2: [payments:write]
        - mtls: []
      x-idempotent: true
      x-rate-limit:
        scope: merchant
        limit: 300
        window: 1s
        burst: 600
        policy: token-bucket
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/Traceparent'
        - $ref: '#/components/parameters/XCorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequest'
      responses:
        '201':
          description: |
            The payment reached a synchronous terminal-for-now state: `AUTHORIZED`,
            `CAPTURED`, `REQUIRES_ACTION`, `FAILED` or `CANCELED`.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            Location: { $ref: '#/components/headers/Location' }
            Idempotent-Replay: { $ref: '#/components/headers/IdempotentReplay' }
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
            RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }
            RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }
            RateLimit-Reset: { $ref: '#/components/headers/RateLimitReset' }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
              examples:
                authorized:
                  summary: Authorized, manual capture
                  value:
                    id: pay_01JB8Z9K2QW3E4R5T6Y7V8J9N0
                    merchantId: mrc_01JB8Z11111111111111111111
                    state: AUTHORIZED
                    amount: { amount: 1050, currency: USD }
                    authorizedAmount: { amount: 1050, currency: USD }
                    capturedAmount: { amount: 0, currency: USD }
                    refundedAmount: { amount: 0, currency: USD }
                    paymentMethod: CARD
                    captureMode: MANUAL
                    riskDecision: ALLOW
                    riskScore: 12
                    threeDsStatus: EXEMPT_LOW_VALUE
                    routingPlanId: rpl_01JB8ZEEEEEEEEEEEEEEEEEEEE
                    currentAttemptId: att_01JB8Z44444444444444444444
                    statementDescriptor: ACME* ORDER 4471
                    metadata: { orderId: '4471' }
                    reconciliationRequired: false
                    expiresAt: '2026-09-02T14:03:11.000Z'
                    version: 3
                    createdAt: '2026-08-26T14:03:10.980Z'
                    updatedAt: '2026-08-26T14:03:11.412Z'
                requiresAction:
                  summary: 3DS challenge required
                  value:
                    id: pay_01JB8ZPPPPPPPPPPPPPPPPPPPP
                    merchantId: mrc_01JB8Z11111111111111111111
                    state: REQUIRES_ACTION
                    amount: { amount: 89900, currency: EUR }
                    capturedAmount: { amount: 0, currency: EUR }
                    refundedAmount: { amount: 0, currency: EUR }
                    paymentMethod: CARD
                    captureMode: AUTOMATIC
                    riskDecision: REQUIRE_3DS
                    riskScore: 61
                    threeDsStatus: CHALLENGE_REQUIRED
                    nextAction:
                      type: REDIRECT
                      redirectUrl: https://3ds.adyen.example/challenge/6f2a1c
                      expiresAt: '2026-08-26T14:18:11.000Z'
                    routingPlanId: rpl_01JB8ZQQQQQQQQQQQQQQQQQQQQ
                    currentAttemptId: att_01JB8ZRRRRRRRRRRRRRRRRRRRR
                    reconciliationRequired: false
                    version: 2
                    createdAt: '2026-08-26T14:03:10.980Z'
                    updatedAt: '2026-08-26T14:03:11.204Z'
        '202':
          description: |
            The outcome is not yet known. The payment is `PROCESSING` (gateway timeout or
            ambiguous transport error, attempt recorded `TIMEOUT_UNKNOWN`) or `PENDING`
            (asynchronous payment method). This is **not** a failure. Poll the payment.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            Location: { $ref: '#/components/headers/Location' }
            Idempotent-Replay: { $ref: '#/components/headers/IdempotentReplay' }
            Retry-After: { $ref: '#/components/headers/RetryAfter' }
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
              examples:
                processingUnknown:
                  summary: Gateway timed out; outcome unknown, reconciliation queued
                  value:
                    id: pay_01JB8ZSSSSSSSSSSSSSSSSSSSS
                    merchantId: mrc_01JB8Z11111111111111111111
                    state: PROCESSING
                    amount: { amount: 1050, currency: USD }
                    capturedAmount: { amount: 0, currency: USD }
                    refundedAmount: { amount: 0, currency: USD }
                    paymentMethod: CARD
                    captureMode: AUTOMATIC
                    riskDecision: ALLOW
                    currentAttemptId: att_01JB8ZHHHHHHHHHHHHHHHHHHHH
                    reconciliationRequired: true
                    version: 2
                    createdAt: '2026-08-26T14:03:10.980Z'
                    updatedAt: '2026-08-26T14:03:19.001Z'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalServerError' }
        '502': { $ref: '#/components/responses/BadGateway' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '504': { $ref: '#/components/responses/GatewayTimeout' }
    get:
      operationId: listPayments
      tags: [Payments]
      summary: List payments
      description: |
        Cursor-paginated payment list, newest first. Reads are served from a replica and may
        be up to 1 s stale; a caller that has just written reads its own write via the
        write-token fallback to the primary.
      security:
        - oauth2: [payments:read]
        - mtls: []
      x-idempotent: true
      x-rate-limit:
        scope: merchant
        limit: 60
        window: 1s
        burst: 120
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/Traceparent'
        - $ref: '#/components/parameters/XCorrelationId'
        - name: merchantId
          in: query
          required: false
          description: |
            Restrict to one merchant. Required for principals whose token is tenant-scoped
            rather than merchant-scoped when the tenant has more than 1000 merchants.
          schema:
            $ref: '#/components/schemas/MerchantId'
          example: mrc_01JB8Z11111111111111111111
        - name: state
          in: query
          required: false
          description: Filter by payment state. Repeatable.
          schema:
            type: array
            items:
              $ref: '#/components/schemas/PaymentState'
          style: form
          explode: true
          example: [AUTHORIZED, CAPTURED]
        - name: createdAfter
          in: query
          required: false
          description: Inclusive lower bound on creation time (RFC 3339, UTC).
          schema:
            type: string
            format: date-time
          example: '2026-08-01T00:00:00.000Z'
        - name: createdBefore
          in: query
          required: false
          description: Exclusive upper bound on creation time (RFC 3339, UTC).
          schema:
            type: string
            format: date-time
          example: '2026-09-01T00:00:00.000Z'
        - name: reconciliationRequired
          in: query
          required: false
          description: Only payments parked awaiting reconciliation.
          schema:
            type: boolean
          example: true
      responses:
        '200':
          description: A page of payments.
          headers:
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentPage'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalServerError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /v1/payments/{paymentId}:
    parameters:
      - $ref: '#/components/parameters/PaymentIdPath'
    get:
      operationId: getPayment
      tags: [Payments]
      summary: Get a payment
      description: |
        Returns the payment with its attempts, refunds and routing plan reference. This is the
        endpoint to poll after a `202`, after a `504`, or while a payment is in
        `REQUIRES_ACTION`.
      security:
        - oauth2: [payments:read]
        - mtls: []
      x-idempotent: true
      x-rate-limit:
        scope: merchant
        limit: 300
        window: 1s
        burst: 600
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/Traceparent'
        - $ref: '#/components/parameters/XCorrelationId'
        - $ref: '#/components/parameters/IfNoneMatch'
        - name: expand
          in: query
          required: false
          description: |
            Embed related resources. `attempts` and `refunds` are embedded by default;
            `routingPlan` is not, because it is an audit artifact and is comparatively large.
          schema:
            type: array
            items:
              type: string
              enum: [attempts, refunds, routingPlan]
          style: form
          explode: true
          example: [routingPlan]
      responses:
        '200':
          description: The payment.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            X-Request-Id: { $ref: '#/components/headers/XRequestId' }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '304':
          description: Not modified.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '500': { $ref: '#/components/responses/InternalServerError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      description: |
        OAuth 2.0 client credentials. Tokens are `ES256` JWTs with a 15-minute TTL and
        audience `https://api.example.com`.

        **The tenant is taken from the token, never from the request.** A `tenantId` in a body
        or query string that disagrees with the token is a security event:
        `403 TENANT_MISMATCH`, audited and alerted.

        The token endpoint is rate limited to 100 requests / 5 min per IP and 20 / min per
        client. Scopes are minted per client, not requested per call: asking for a scope the
        client was not granted returns `invalid_scope` rather than a narrowed token.
      flows:
        clientCredentials:
          tokenUrl: https://auth.example.com/oauth2/token
          refreshUrl: https://auth.example.com/oauth2/token
          scopes:
            merchants:read: Read merchants, business profiles, bank accounts and principals.
            merchants:write: Create and update merchants.
            merchants:suspend: Suspend and resume a merchant.
            merchants:terminate: Terminate a merchant. Dual control required.
            onboarding:read: Read onboarding cases and step history.
            onboarding:write: Start and drive onboarding.
            onboarding:approve: Signal a manual onboarding gate. Dual control required.
            config:read: Read configuration documents and version history.
            config:write: Publish a new configuration version.
            config:rollback: Roll back configuration to an earlier version. Dual control required.
            gateways:read: Read the gateway catalogue and health.
            gateways:write: Register and deprecate gateways. Platform administrators only.
            credentials:rotate: Rotate gateway credentials. Dual control required.
            credentials:read: Read credential metadata for the caller's own path. Internal workloads only.
            payments:read: Read payments, attempts and refunds.
            payments:write: Create payments. Machine clients only - no human role holds this scope.
            payments:capture: Capture an authorized payment.
            payments:refund: Refund a captured payment. Dual control for human principals.
            payments:void: Void an authorization.
            payments:replay_dlq: Replay dead-lettered payment events. Operators only.
            ledger:read: Read ledger accounts and entries.
            ledger:write: Append ledger entries. Event consumer only.
...
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Client-supplied opaque key scoping one logical operation to one effect. Required on
        every unsafe operation.

        The idempotency scope is
        `(tenant_id, merchant_id, method, path_template, idempotency_key)`. A SHA-256
        fingerprint of the JCS-canonicalized request body is stored with the key: the same key
        with a different body is `422 IDEMPOTENCY_KEY_REUSED`, which is the check that catches
        a client reusing one key for two different payments.

        A concurrent duplicate while the first request is still in flight is
        `409 IDEMPOTENT_REQUEST_IN_PROGRESS` with `Retry-After: 1`. The platform does not
        block the second caller and does not process twice.

        Recommended: a UUIDv4 per logical operation, generated **before** the first attempt and
        reused across every retry of that operation. Records are retained 7 days.
      schema:
        type: string
        minLength: 1
        maxLength: 255
        pattern: '^[\x21-\x7E]{1,255}$'
      example: 6b3f1a52-9e7c-4b4f-8f4a-1c2d3e4f5a6b
...
    XRequestId:
      name: X-Request-Id
      in: header
      required: false
      description: |
        Caller-supplied request identifier, echoed on every response including errors. If
        absent the platform generates a `req_`-prefixed ULID. It appears in the problem
        document as `requestId` and on every log line and audit record for this request.
      schema:
        type: string
        maxLength: 128
      example: req_01JB8Z22222222222222222222
    Traceparent:
      name: traceparent
      in: header
      required: false
      description: |
        W3C Trace Context. Propagated into every span, event envelope and gateway call, and
        echoed on the response. If absent a new trace is started. Format
        `00-<32 hex trace-id>-<16 hex span-id>-<2 hex flags>`.
      schema:
        type: string
        pattern: '^[0-9a-f]{2}-[0-9a-f]{32}-[0-9a-f]{16}-[0-9a-f]{2}$'
      example: 00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01
    XCorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      description: |
        Business correlation identifier that stays constant across an entire causal fan-out,
        so one API call's full consequences - payments, attempts, events, audit records - are
        one query. Copied into `correlationid` on every event this request produces. Defaults
        to the request id.
      schema:
        type: string
        maxLength: 128
      example: req_01JB8Z22222222222222222222
    Limit:
      name: limit
      in: query
      required: false
      description: Maximum items in the page.
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 200
        default: 50
      example: 50
...
  headers:
    ETag:
      description: |
        The resource's optimistic-concurrency token: the aggregate version for merchants and
        payments, the document digest for configuration. Echo it in `If-Match`.
      required: true
      schema:
        type: string
      example: '"7"'
    Location:
      description: Absolute URL of the created or affected resource.
      required: false
      schema:
        type: string
        format: uri
      example: https://api.example.com/v1/payments/pay_01JB8Z9K2QW3E4R5T6Y7V8J9N0
    IdempotentReplay:
      description: |
        `true` when this response was replayed from a stored idempotency record rather than
        produced by executing the operation again. Absent or `false` on first execution.
      required: false
      schema:
        type: boolean
      example: true
    XRequestId:
      description: The request identifier, echoed or generated. Always present.
      required: true
      schema:
        type: string
      example: req_01JB8Z22222222222222222222
    RetryAfter:
      description: |
        Seconds to wait before retrying. Present on `409 IDEMPOTENT_REQUEST_IN_PROGRESS`,
        `429`, `503` and on `202` responses where polling is the correct next step.
      required: false
      schema:
        type: integer
        format: int32
        minimum: 0
      example: 1
    RateLimitLimit:
      description: Requests permitted in the current window for this scope.
      required: false
      schema:
        type: integer
        format: int32
      example: 300
    RateLimitRemaining:
      description: Requests remaining in the current window.
      required: false
      schema:
        type: integer
        format: int32
      example: 271
    RateLimitReset:
      description: Seconds until the current window resets.
      required: false
      schema:
        type: integer
        format: int32
      example: 1
...
  responses:
    BadRequest:
      description: |
        The request could not be parsed or failed L1 schema validation. Not retryable without
        changing the request.
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            validationFailed:
              summary: VALIDATION_FAILED
              value:
                type: https://errors.example.com/VALIDATION_FAILED
                title: Request failed schema validation
                status: 400
                code: VALIDATION_FAILED
                detail: '2 fields failed validation.'
                category: VALIDATION
                retryable: false
                requestId: req_01JB8Z22222222222222222222
                traceId: 4bf92f3577b34da6a3ce929d0e0e4736
                details:
                  - field: amount.currency
                    code: NOT_ISO_4217
                    message: '"US" is not a valid ISO 4217 alphabetic code.'
                  - field: statementDescriptor
                    code: TOO_LONG
                    message: Must be at most 22 characters.
                docsUrl: https://docs.example.com/errors/VALIDATION_FAILED
            sensitiveDataInRequest:
              summary: SENSITIVE_DATA_IN_REQUEST
              description: |
                The L1 PAN detector found a Luhn-valid 13-19 digit sequence. The offending
                value is neither echoed nor logged, and a security event is raised. This API
                accepts gateway tokens and vault references only.
              value:
                type: https://errors.example.com/SENSITIVE_DATA_IN_REQUEST
                title: Request contained cardholder data
                status: 400
                code: SENSITIVE_DATA_IN_REQUEST
                detail: >-
                  A field in this request matched the primary account number detector. The
                  value has been discarded and was not logged. Submit a gateway token or a
                  vault reference instead of card data.
                category: VALIDATION
                retryable: false
                requestId: req_01JB8Z22222222222222222222
                traceId: 4bf92f3577b34da6a3ce929d0e0e4736
                details:
                  - field: paymentMethod.token
                    code: PAN_DETECTED
                    message: Value redacted. Cardholder data must never reach this API.
                docsUrl: https://docs.example.com/errors/SENSITIVE_DATA_IN_REQUEST
            idempotencyKeyRequired:
              summary: IDEMPOTENCY_KEY_REQUIRED
              value:
                type: https://errors.example.com/IDEMPOTENCY_KEY_REQUIRED
                title: Idempotency-Key header is required
                status: 400
                code: IDEMPOTENCY_KEY_REQUIRED
                detail: 'POST /v1/payments requires an Idempotency-Key header of 1-255 characters.'
                category: VALIDATION
                retryable: false
                requestId: req_01JB8Z22222222222222222222
                traceId: 4bf92f3577b34da6a3ce929d0e0e4736
                docsUrl: https://docs.example.com/errors/IDEMPOTENCY_KEY_REQUIRED
    Unauthorized:
      description: Authentication is missing, malformed, expired or failed signature verification.
      headers:
        X-Request-Id: { $ref: '#/components/headers/XRequestId' }
        WWW-Authenticate:
          description: Challenge indicating the expected scheme.
          schema:
            type: string
          example: Bearer realm="payments-platform", error="invalid_token"
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            unauthenticated:
              summary: UNAUTHENTICATED
              value:
                type: https://errors.example.com/UNAUTHENTICATED
                title: Authentication failed
                status: 401
                code: UNAUTHENTICATED
                detail: The access token has expired.
                category: AUTHENTICATION
                retryable: false
                requestId: req_01JB8Z22222222222222222222
                traceId: 4bf92f3577b34da6a3ce929d0e0e4736
                docsUrl: https://docs.example.com/errors/UNAUTHENTICATED
            webhookSignatureInvalid:
              summary: WEBHOOK_SIGNATURE_INVALID
              value:
                type: https://errors.example.com/WEBHOOK_SIGNATURE_INVALID
                title: Webhook signature verification failed
                status: 401
                code: WEBHOOK_SIGNATURE_INVALID
                detail: The X-Gateway-Signature header did not verify over the received body.
                category: AUTHENTICATION
                retryable: false
                requestId: req_01JB8Z22222222222222222222
                traceId: 4bf92f3577b34da6a3ce929d0e0e4736
                docsUrl: https://docs.example.com/errors/WEBHOOK_SIGNATURE_INVALID
...
  schemas:
...
    CreatePaymentRequest:
      type: object
      title: Create payment request
      required: [merchantId, amount, paymentMethod, paymentMethodReference]
      additionalProperties: false
      properties:
        merchantId:
          allOf:
            - $ref: '#/components/schemas/MerchantId'
          description: |
            The merchant on whose behalf the payment is made. Must belong to the token's
            tenant and be `ACTIVE`; anything else is `409 MERCHANT_NOT_ACTIVE`.
        amount:
          allOf:
            - $ref: '#/components/schemas/PositiveMoney'
          description: |
            Minor units and ISO 4217 code. Must be strictly positive, within the merchant's
            configured maximum and in a supported currency.
        paymentMethod:
          $ref: '#/components/schemas/PaymentMethod'
        paymentMethodReference:
          $ref: '#/components/schemas/PaymentMethodReference'
        captureMode:
          allOf:
            - $ref: '#/components/schemas/CaptureMode'
          default: AUTOMATIC
          description: |
            `AUTOMATIC` captures on authorization. `MANUAL` leaves the payment `AUTHORIZED`
            until an explicit capture.
        statementDescriptor:
          type: string
          maxLength: 22
          pattern: '^[A-Za-z0-9 *.,\-]{5,22}$'
          description: |
            What the payer sees on their statement. A recognisable descriptor is the cheapest
            chargeback prevention there is.
          example: ACME* ORDER 4471
        payerCountry:
          allOf:
            - $ref: '#/components/schemas/CountryCode'
          description: |
            Payer's country, for risk and routing. Checked against the merchant's `countries`
            and `risk.blockedCountries`.
        reference:
          type: string
          maxLength: 128
          description: |
            The merchant's own order reference. Distinct from the idempotency key: this is
            business data, the key is a protocol concern.
          example: order-4471
        metadata:
          type: object
          additionalProperties:
            type: string
            maxLength: 512
          maxProperties: 50
          description: |
            Arbitrary merchant key-value data, echoed on reads. Do **not** put cardholder
            data, credentials or personal data here - the L1 PAN detector inspects these
            values like every other string.
          example:
            orderId: '4471'
        threeDsPreference:
          type: string
          enum: [AUTOMATIC, FORCE_CHALLENGE, REQUEST_EXEMPTION]
          default: AUTOMATIC
          description: |
            A *request*, not an instruction. The risk policy decides; `FORCE_CHALLENGE` can
            escalate, and `REQUEST_EXEMPTION` is honoured only when the payment actually
            qualifies for one.
          example: AUTOMATIC
        preferredGateway:
          allOf:
            - $ref: '#/components/schemas/GatewayCode'
          description: |
            A routing hint for contractual pinning. Honoured only when the gateway is
            eligible, certified and healthy; otherwise routing proceeds normally and the plan
            records why the preference was not taken.
      example:
        merchantId: mrc_01JB8Z11111111111111111111
        amount: { amount: 1050, currency: USD }
        paymentMethod: CARD
        paymentMethodReference:
          type: GATEWAY_TOKEN
          gatewayCode: stripe
          token: pm_1OqZ4L2eZvKYlo2CqXbTgHkF
          brand: VISA
          last4: '4242'
          expiryMonth: 11
          expiryYear: 2029
        captureMode: MANUAL
        statementDescriptor: ACME* ORDER 4471
        payerCountry: US
        reference: order-4471
        metadata: { orderId: '4471' }
        threeDsPreference: AUTOMATIC
...